+++ lib/flip_layout.dart
@@ -2,3 +2,4 @@ export 'src/layout_motion.dart';
 export 'src/animated_layout.dart';
 export 'src/motion_group.dart';
 export 'src/motion_config.dart';
+export 'src/spring_curve.dart';